+++ api/parsers/tlogparser.py
@@ -1,12 +1,12 @@
 import pandas as pd
 from cesium_entity import CesiumEntity
-from parser import Parser
+from .parser import Parser
 from pymavlink import mavutil
 
-class ArduParser(Parser):
+class TLOGParser(Parser):
     def __init__(self):
         super().__init__()
-        self.name = "ardu_parser"
+        self.name = "tlog_parser"
         self.initDefaultEntity()
         self.initEntities()
 
@@ -14,13 +14,13 @@ class ArduParser(Parser):
         mlog = mavutil.mavlink_connection(filename)
         buf = {}
         
-        while mlog.remaining:
+        while True:
             m = mlog.recv_match()
             if (m is None): break
             name = m.get_type()
             data = m.to_dict()
-            if 'TimeUS' in list(data.keys()):
-                data['timestamp_tiplot'] = data['TimeUS'] / 1e6
+            if 'time_boot_ms' in list(data.keys()):
+                data['timestamp_tiplot'] = data['time_boot_ms'] / 1e3
             else:
                 #ignore tables with no timestamp
                 continue
@@ -38,23 +38,22 @@ class ArduParser(Parser):
 
 
     def initDefaultEntity(self):
-        self.default_entity = CesiumEntity(name='ardu pilot default entity',
-                              useRPY=False,
-                              useXYZ=False,
+        self.default_entity = CesiumEntity(name='tlog default entity',
+                              useRPY=True,
+                              useXYZ=True,
                               color="#ffffff",
                               pathColor="#0000ff",
                               position={
-                                  'table':'AHR2',
-                                  'longitude': 'Lng',
-                                  'lattitude': 'Lat',
-                                  'altitude': 'Alt',
+                                  'table':'LOCAL_POSITION_NED',
+                                  'x': 'x',
+                                  'y': 'y',
+                                  'z': 'z',
                               },
                               attitude={
-                                  'table':'AHR2',
-                                  'q0':'Q1',
-                                  'q1':'Q2',
-                                  'q2':'Q3',
-                                  'q3':'Q4',
+                                  'table':'ATTITUDE',
+                                  'roll':'roll',
+                                  'pitch':'pitch',
+                                  'yaw':'yaw',
                               })
 
     def setDefaultEntities(self):
